Refactor property ID type to Long and update email service to improve…

+++ src/main/java/com/javaweb/model/PropertyDTO.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public class PropertyDTO {
 
-    private Integer propertyID;
+    private Long propertyID;
     private String addressLine1;
     private String addressLine2;
     private String region;
@@ -34,11 +34,11 @@ public class PropertyDTO {
 
 
     // Getters and Setters
-    public Integer getPropertyID() {
+    public Long getPropertyID() {
         return propertyID;
     }
 
-    public void setPropertyID(Integer propertyID) {
+    public void setPropertyID(Long propertyID) {
         this.propertyID = propertyID;
     }
